Updating tooling and fixing lint errors
