+++ .eslintrc
@@ -1,0 +1,4 @@
+{
+  "extends": "vtex",
+  "root": true
+}
