+++ react/hooks/useProfileAllowed.ts
@@ -1,4 +1,5 @@
 import { useQuery } from 'react-apollo'
+
 import checkProfileAllowedQuery from '../graphql/documents.graphql'
 
 export const useProfileAllowed = (skip: boolean) => {
