+++ react/package.json
@@ -1,6 +1,5 @@
 {
   "scripts": {
-    "lint": "tsc --noEmit && eslint --quiet --fix --ext ts,tsx .",
     "test": "vtex-test-tools test"
   },
   "dependencies": {
@@ -17,15 +16,12 @@
     "@types/jest": "^24.0.21",
     "@types/node": "^12.12.3",
     "@types/ramda": "^0.26.33",
-    "@types/react": "^16.8.22",
-    "@types/react-intl": "^2.3.18",
     "@vtex/test-tools": "^0.3.2",
     "@vtex/tsconfig": "^0.2.0",
-    "eslint": "5.x",
-    "eslint-config-vtex-react": "^5.0.1",
-    "prettier": "^1.18.2",
-    "prop-types": "^15.7.2",
-    "typescript": "3.8.3"
+    "vtex.css-handles": "http://vtex.vtexassets.com/_v/public/typings/v1/vtex.css-handles@0.4.4/public/@types/vtex.css-handles",
+    "vtex.render-runtime": "http://vtex.vtexassets.com/_v/public/typings/v1/vtex.render-runtime@8.126.11/public/@types/vtex.render-runtime",
+    "vtex.store": "http://vtex.vtexassets.com/_v/public/typings/v1/vtex.store@2.111.1/public/@types/vtex.store",
+    "vtex.store-graphql": "http://vtex.vtexassets.com/_v/public/typings/v1/vtex.store-graphql@2.138.0/public/@types/vtex.store-graphql"
   },
   "version": "0.1.1"
 }
