+++ react/typings/graphql.d.ts
@@ -1,5 +1,5 @@
 declare module '*.graphql' {
-  import { DocumentNode } from 'graphql'
+  import type { DocumentNode } from 'graphql'
 
   const value: DocumentNode
   export default value
