--- react/typings/vtex.css-handles.d.ts
@@ -1,1 +1,0 @@
-declare module 'vtex.css-handles'
